+++ packages/lavas-core-vue/core/builder/base-builder.js
@@ -12,8 +12,8 @@ import SkeletonWebpackPlugin from 'vue-skeleton-webpack-plugin';
 import VueSSRClientPlugin from 'vue-server-renderer/client-plugin';
 import OmmitCSSPlugin from '../plugins/ommit-css-webpack-plugin';
 
-import {TEMPLATE_HTML, DEFAULT_ENTRY_NAME, DEFAULT_SKELETON_PATH, CONFIG_FILE,
-    LAVAS_DIRNAME_IN_DIST, CLIENT_MANIFEST, STORE_FILE} from '../constants';
+import {TEMPLATE_HTML, SPA_TEMPLATE_HTML, DEFAULT_ENTRY_NAME, DEFAULT_SKELETON_PATH,
+    CONFIG_FILE, LAVAS_DIRNAME_IN_DIST, CLIENT_MANIFEST, STORE_FILE} from '../constants';
 import {assetsPath, resolveAliasPath, camelCaseToDash} from '../utils/path';
 import {enableHotReload} from '../utils/webpack';
 import * as JsonUtil from '../utils/json';
@@ -139,21 +139,6 @@ export default class BaseBuilder {
         );
     }
 
-    async writeLavasLink() {
-        let lavasLinkTemplate = await readFile(this.templatesPath('LavasLink.js.tmpl'), 'utf8');
-        await this.writeFileToLavasDir('LavasLink.js', template(lavasLinkTemplate)({
-            entryConfig: JsonUtil.stringify(this.config.entries.map(entry => {
-                // only select necessary keys
-                return {
-                    name: entry.name,
-                    urlReg: entry.urlReg
-                };
-            })),
-            base: this.config.router.base,
-            mode: this.config.router.mode
-        }));
-    }
-
     /**
      * write an entry file for skeleton components
      *
@@ -174,30 +159,29 @@ export default class BaseBuilder {
      *
      * @param {Object} spaConfig spaConfig
      * @param {string} baseUrl baseUrl from config/router
-     * @param {?string} entryName entry name in MPA, undefined in SPA
      * @return {string} resolvedTemplatePath html template's path
      */
-    async addHtmlPlugin(spaConfig, baseUrl = '/', entryName) {
+    async addHtmlPlugin(spaConfig, baseUrl = '/') {
         // allow user to provide a custom HTML template
         let {globals: {rootDir}, skeleton: {enable: enableSkeleton}, build: {cssExtract}} = this.config;
         let htmlFilename;
         let templatePath;
         let tempTemplatePath;
 
-        if (entryName) {
-            htmlFilename = `${entryName}/${entryName}.html`;
-            templatePath = join(rootDir, `entries/${entryName}/${TEMPLATE_HTML}`);
-            tempTemplatePath = `${entryName}/${TEMPLATE_HTML}`;
-        }
-        else {
-            htmlFilename = `${DEFAULT_ENTRY_NAME}.html`;
+        htmlFilename = `${DEFAULT_ENTRY_NAME}.html`;
+
+        // find core/spa.html.tmpl
+        templatePath = join(rootDir, `core/${SPA_TEMPLATE_HTML}`);
+        if (!await pathExists(templatePath)) {
+            // find core/index.html.tmpl
             templatePath = join(rootDir, `core/${TEMPLATE_HTML}`);
-            tempTemplatePath = TEMPLATE_HTML;
         }
 
         if (!await pathExists(templatePath)) {
-            throw new Error(`${TEMPLATE_HTML} required for entry: ${entryName || DEFAULT_ENTRY_NAME}`);
+            throw new Error(`${SPA_TEMPLATE_HTML} or ${TEMPLATE_HTML} required`);
         }
+
+        tempTemplatePath = basename(templatePath);
 
         // write HTML template used by html-webpack-plugin which doesn't support template STRING
         let resolvedTemplatePath = await this.writeFileToLavasDir(
@@ -229,11 +213,11 @@ export default class BaseBuilder {
             favicon: assetsPath('img/icons/favicon.ico'),
             chunksSortMode: 'dependency',
             cache: false,
-            chunks: ['manifest', 'vue', 'vendor', entryName || DEFAULT_ENTRY_NAME],
+            chunks: ['manifest', 'vue', 'vendor', DEFAULT_ENTRY_NAME],
             config: this.config // use config in template
         }]);
 
-        return resolvedTemplatePath;
+        return {resolvedTemplatePath, tempTemplatePath};
     }
 
     /**
@@ -365,7 +349,11 @@ export default class BaseBuilder {
                         .entry(DEFAULT_ENTRY_NAME).add('./core/entry-client.js');
 
                     // add html-webpack-plugin
-                    let customTemplatePath = await this.addHtmlPlugin(clientConfig, router.base);
+                    let {
+                        resolvedTemplatePath: customTemplatePath,
+                        tempTemplatePath
+                    } = await this.addHtmlPlugin(clientConfig, router.base);
+                    // let customTemplatePath = await this.addHtmlPlugin(clientConfig, router.base);
 
                     // add vue-skeleton-webpack-plugin
                     if (skeleton && skeleton.enable) {
@@ -377,7 +365,7 @@ export default class BaseBuilder {
                         // watch html
                         this.addWatcher(customTemplatePath, 'change', async () => {
                             await this.writeFileToLavasDir(
-                                TEMPLATE_HTML,
+                                tempTemplatePath,
                                 templateUtil.client(await readFile(customTemplatePath, 'utf8'), router.base)
                             );
                         });